+++ aklearn/logisticreg.py
@@ -148,7 +148,7 @@ class Logistic(Classification):
                                             coefficient)
             return value
 
-        dimension = np.ones(labels.shape[1])
+        dimension = np.ones(labels.shape[1], dtype = np.int8)
         optimum = _minimize_bfgs(negative_log_likelihood, dimension)
         mle = optimum.x  # Obtain argmin
         return mle
